requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.9.24
cffi==1.15.1
channels==2.4.0
channels-redis==3.0.1
charset-normalizer==2.1.1
colorama==0.4.5
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==2.5.0
distlib==0.3.6
Django==4.1
djangorestframework==3.14.0
drf-yasg==1.21.4
filelock==3.8.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
zope.interface==5.4.0
